refactor: moving sharding to precompute
